refactor(blockchain-link): websocket unification
