+++ packages/blockchain-link/src/workers/blockbook/websocket.ts
@@ -1,6 +1,4 @@
 import WebSocket from 'ws';
-import { createDeferred, Deferred } from '@trezor/utils/lib/createDeferred';
-import { TypedEmitter } from '@trezor/utils/lib/typedEventEmitter';
 
 import { CustomError } from '@trezor/blockchain-link-types/lib/constants/errors';
 import type {
@@ -21,183 +19,24 @@ import type {
     AccountBalanceHistoryParams,
 } from '@trezor/blockchain-link-types/lib/params';
 
-interface Subscription {
-    id: string;
-    type: 'notification' | 'block' | 'mempool' | 'fiatRates';
-    callback: (result: any) => void;
-}
-
-interface Options {
-    url: string;
-    timeout?: number;
-    pingTimeout?: number;
-    keepAlive?: boolean;
-    agent?: WebSocket.ClientOptions['agent'];
-    headers?: WebSocket.ClientOptions['headers'];
-}
-
-const DEFAULT_TIMEOUT = 20 * 1000;
-const DEFAULT_PING_TIMEOUT = 50 * 1000;
+import { BaseWebsocket } from '../baseWebsocket';
 
 interface BlockbookEvents {
     block: BlockNotification;
     mempool: MempoolTransactionNotification;
     notification: AddressNotification;
     fiatRates: FiatRatesNotification;
-    error: string;
-    disconnected: undefined;
 }
 
-export class BlockbookAPI extends TypedEmitter<BlockbookEvents> {
-    options: Options;
-    ws: WebSocket | undefined;
-    messageID = 0;
-    messages: Deferred<any>[] = [];
-    subscriptions: Subscription[] = [];
-    pingTimeout: ReturnType<typeof setTimeout> | undefined;
-    connectionTimeout: ReturnType<typeof setTimeout> | undefined;
-
-    constructor(options: Options) {
-        super();
-        this.options = options;
-    }
-
-    setConnectionTimeout() {
-        this.clearConnectionTimeout();
-        this.connectionTimeout = setTimeout(
-            this.onTimeout.bind(this),
-            this.options.timeout || DEFAULT_TIMEOUT,
-        );
-    }
-
-    clearConnectionTimeout() {
-        if (this.connectionTimeout) {
-            clearTimeout(this.connectionTimeout);
-            this.connectionTimeout = undefined;
-        }
-    }
-
-    setPingTimeout() {
-        if (this.pingTimeout) {
-            clearTimeout(this.pingTimeout);
-        }
-        this.pingTimeout = setTimeout(
-            this.onPing.bind(this),
-            this.options.pingTimeout || DEFAULT_PING_TIMEOUT,
-        );
-    }
-
-    private rejectAllPending(code: string, message?: string) {
-        this.messages.forEach(m => m.reject(new CustomError(code, message)));
-        this.messages = [];
-    }
-
-    onTimeout() {
-        const { ws } = this;
-        if (!ws) return;
-        if (ws.listenerCount('open') > 0) {
-            ws.emit('error', new CustomError('websocket_timeout'));
-            try {
-                ws.close();
-            } catch (error) {
-                // empty
-            }
-        } else {
-            this.rejectAllPending('websocket_timeout');
-            ws.close();
-        }
-    }
-
-    async onPing() {
-        // make sure that connection is alive if there are subscriptions
-        if (this.ws && this.isConnected()) {
-            try {
-                if (this.subscriptions.length > 0 || this.options.keepAlive) {
-                    await this.getBlockHash(1);
-                } else {
-                    this.ws.close();
-                }
-            } catch (error) {
-                // empty
-            }
-        }
-    }
-
-    onError() {
-        this.onClose();
-    }
-
-    send: Send = (method, params = {}) => {
-        const { ws } = this;
-        if (!ws) throw new CustomError('websocket_not_initialized');
-        const id = this.messageID.toString();
-
-        const dfd = createDeferred(id);
-        const req = {
-            id,
-            method,
-            params,
-        };
-
-        this.messageID++;
-        this.messages.push(dfd);
-
-        this.setConnectionTimeout();
-        this.setPingTimeout();
-
-        ws.send(JSON.stringify(req));
-        return dfd.promise as Promise<any>;
-    };
-
-    onmessage(message: string) {
-        try {
-            const resp = JSON.parse(message);
-            const { id, data } = resp;
-            const dfd = this.messages.find(m => m.id === id);
-            if (dfd) {
-                if (data.error) {
-                    dfd.reject(new CustomError('websocket_error_message', data.error.message));
-                } else {
-                    dfd.resolve(data);
-                }
-                this.messages.splice(this.messages.indexOf(dfd), 1);
-            } else {
-                const subs = this.subscriptions.find(s => s && s.id === id);
-                if (subs) {
-                    subs.callback(data);
-                }
-            }
-        } catch (error) {
-            // empty
-        }
-
-        if (this.messages.length === 0) {
-            this.clearConnectionTimeout();
-        }
-        this.setPingTimeout();
-    }
-
-    private connectPromise: Promise<void> | undefined;
-
-    async connect() {
-        // if connecting already, just return the promise
-        if (this.connectPromise) {
-            return this.connectPromise;
-        }
-
-        if (this.ws?.readyState === WebSocket.CLOSING) {
-            await new Promise<void>(resolve => this.once('disconnected', resolve));
-        }
-
+export class BlockbookAPI extends BaseWebsocket<BlockbookEvents> {
+    protected createWebsocket() {
         // url validation
         let { url } = this.options;
         if (typeof url !== 'string') {
             throw new CustomError('websocket_no_url');
         }
 
-        if (url.startsWith('https')) {
-            url = url.replace('https', 'wss');
-        }
+        // this resolves https->wss as well
         if (url.startsWith('http')) {
             url = url.replace('http', 'ws');
         }
@@ -206,17 +45,9 @@ export class BlockbookAPI extends TypedEmitter<BlockbookEvents> {
             url += suffix;
         }
 
-        // set connection timeout before WebSocket initialization
-        // it will be be cancelled by this.init or this.dispose after the error
-        this.setConnectionTimeout();
-
-        // create deferred promise
-        const dfd = createDeferred<void>(-1);
-        this.connectPromise = dfd.promise;
-
         // initialize connection,
         // options are not used in web builds (see ./src/utils/ws)
-        const ws = new WebSocket(url, {
+        return new WebSocket(url, {
             agent: this.options.agent,
             headers: {
                 Origin: 'https://node.trezor.io',
@@ -224,48 +55,13 @@ export class BlockbookAPI extends TypedEmitter<BlockbookEvents> {
                 ...this.options.headers,
             },
         });
-        ws.once('error', error => {
-            this.onClose();
-            dfd.reject(new CustomError('websocket_runtime_error', error.message));
-        });
-        ws.on('open', () => {
-            this.init();
-            dfd.resolve();
-        });
-
-        this.ws = ws;
-
-        // wait for onopen event
-        return dfd.promise.finally(() => {
-            this.connectPromise = undefined;
-        });
     }
 
-    init() {
-        const { ws } = this;
-        if (!ws || !this.isConnected()) {
-            throw Error('Blockbook websocket init cannot be called');
-        }
-        // clear timeout from this.connect
-        this.clearConnectionTimeout();
-
-        // remove previous listeners and add new listeners
-        ws.removeAllListeners();
-        ws.on('error', this.onError.bind(this));
-        ws.on('message', this.onmessage.bind(this));
-        ws.on('close', () => {
-            this.onClose();
-            this.emit('disconnected');
-        });
+    protected ping() {
+        return this.getBlockHash(1);
     }
 
-    disconnect() {
-        this.ws?.close();
-    }
-
-    isConnected() {
-        return this.ws?.readyState === WebSocket.OPEN;
-    }
+    send: Send = (method, params = {}) => this.sendMessage({ method, params });
 
     getServerInfo() {
         return this.send('getInfo');
@@ -319,126 +115,47 @@ export class BlockbookAPI extends TypedEmitter<BlockbookEvents> {
         return this.send('getFiatRatesTickersList', payload);
     }
 
-    private removeSubscription(type: Subscription['type']) {
-        const index = this.subscriptions.findIndex(s => s.type === type);
-        if (index >= 0) {
-            // remove previous subscriptions
-            this.subscriptions.splice(index, 1);
-        }
-        return index;
-    }
-
     subscribeAddresses(addresses: string[]) {
         this.removeSubscription('notification');
-
-        // add new subscription
-        const id = this.messageID.toString();
-        this.subscriptions.push({
-            id,
-            type: 'notification',
-            callback: (result: AddressNotification) => {
-                this.emit('notification', result);
-            },
-        });
+        this.addSubscription('notification', result => this.emit('notification', result));
         return this.send('subscribeAddresses', { addresses });
     }
 
     unsubscribeAddresses() {
         const index = this.removeSubscription('notification');
-        if (index >= 0) {
-            return this.send('unsubscribeAddresses');
-        }
-        return { subscribed: false };
+        return index >= 0 ? this.send('unsubscribeAddresses') : { subscribed: false };
     }
 
     subscribeBlock() {
         this.removeSubscription('block');
-
-        // add new subscription
-        const id = this.messageID.toString();
-        this.subscriptions.push({
-            id,
-            type: 'block',
-            callback: (result: BlockNotification) => {
-                this.emit('block', result);
-            },
-        });
+        this.addSubscription('block', result => this.emit('block', result));
         return this.send('subscribeNewBlock');
     }
 
     unsubscribeBlock() {
         const index = this.removeSubscription('block');
-
-        if (index >= 0) {
-            return this.send('unsubscribeNewBlock');
-        }
-        return { subscribed: false };
+        return index >= 0 ? this.send('unsubscribeNewBlock') : { subscribed: false };
     }
 
     subscribeFiatRates(currency?: string) {
         this.removeSubscription('fiatRates');
-
-        // add new subscription
-        const id = this.messageID.toString();
-        this.subscriptions.push({
-            id,
-            type: 'fiatRates',
-            callback: (result: FiatRatesNotification) => {
-                this.emit('fiatRates', result);
-            },
-        });
+        this.addSubscription('fiatRates', result => this.emit('fiatRates', result));
         return this.send('subscribeFiatRates', { currency });
     }
 
     unsubscribeFiatRates() {
         const index = this.removeSubscription('fiatRates');
-        if (index >= 0) {
-            return this.send('unsubscribeFiatRates');
-        }
-        return { subscribed: false };
+        return index >= 0 ? this.send('unsubscribeFiatRates') : { subscribed: false };
     }
 
     subscribeMempool() {
         this.removeSubscription('mempool');
-
-        // add new subscription
-        const id = this.messageID.toString();
-        this.subscriptions.push({
-            id,
-            type: 'mempool',
-            callback: result => {
-                this.emit('mempool', result);
-            },
-        });
+        this.addSubscription('mempool', result => this.emit('mempool', result));
         return this.send('subscribeNewTransaction');
     }
 
     unsubscribeMempool() {
         const index = this.removeSubscription('mempool');
-        if (index >= 0) {
-            return this.send('unsubscribeNewTransaction');
-        }
-        return { subscribed: false };
-    }
-
-    private onClose() {
-        if (this.pingTimeout) {
-            clearTimeout(this.pingTimeout);
-        }
-        if (this.connectionTimeout) {
-            clearTimeout(this.connectionTimeout);
-        }
-
-        if (this.isConnected()) {
-            this.disconnect();
-        }
-        this.ws?.removeAllListeners();
-
-        this.rejectAllPending('websocket_runtime_error', 'Websocket closed unexpectedly');
-    }
-
-    dispose() {
-        this.onClose();
-        this.removeAllListeners();
+        return index >= 0 ? this.send('unsubscribeNewTransaction') : { subscribed: false };
     }
 }
